LOTS
Populated most scripts to at least get them to the right page (Even though it is not yet fully automated) Tied the scripts to the to .mp3 files and  barcodeImages files for ease of changing locations.  Added variables to the subScripts to easily change ChromeScaling, the Snipe IT URL for different environments, and general Title and Menu Number use for some of the scripts.